add reduce.hpp header
